handle api with category name

+++ lib/features/home/data/models/placeCard_model.dart
@@ -24,9 +24,8 @@ class PlaceCardModel {
       imageUre: json['imageUre'],
       title: json['name'],
       city: json['location']['city'],
-
       shortDescription: json['description'],
-      rating: json['average_rating'],
+      rating: json['average_rating'].toDouble(),
     );
   }
 }
